+++ src/main/java/wooteco/subway/controller/LineController.java
@@ -1,9 +1,13 @@
-package wooteco.subway.line;
+package wooteco.subway.controller;
 
 import org.springframework.dao.DataAccessException;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import wooteco.subway.domain.Line;
+import wooteco.subway.dao.LineDao;
+import wooteco.subway.dto.line.LineRequest;
+import wooteco.subway.dto.line.LineResponse;
 
 import java.net.URI;
 import java.util.List;
