+++ src/main/java/wooteco/subway/dao/StationDao.java
@@ -1,9 +1,10 @@
-package wooteco.subway.station;
+package wooteco.subway.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import wooteco.subway.domain.Station;
 
 import java.sql.PreparedStatement;
 import java.util.List;
@@ -18,7 +19,6 @@ public class StationDao {
     }
 
     public Long save(Station station) {
-//        validateStation(station);
         KeyHolder keyHolder = new GeneratedKeyHolder();
         String sql = "INSERT INTO station (name) VALUES (?)";
 
@@ -30,12 +30,6 @@ public class StationDao {
 
         return keyHolder.getKey().longValue();
     }
-
-//    private void validateStation(Station newStation) {
-//        if (duplicatedNameExists(newStation.getName())) {
-//            throw new IllegalArgumentException("중복된 지하철 역입니다.");
-//        }
-//    }
 
     public List<Station> findAll() {
         String sql = "SELECT * FROM station";
